backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
requests
web3
httpx
SQLAlchemy